視点の調整とobsの型の変更

+++ gym_cable/envs/robot_env.py
@@ -81,9 +81,9 @@ class BaseRobotEnv(GoalEnv):
         self.action_space = spaces.Box(-1.0, 1.0, shape=(n_actions,), dtype="float32")
         self.observation_space = spaces.Dict(
             dict(
-                observation=spaces.Box(-np.inf, np.inf, shape=obs["observation"].shape, dtype="float32"),
+                observation=spaces.Box(-np.inf, np.inf, shape=obs["observation"].shape, dtype="float64"),
                 rgb_image=spaces.Box(0, 255, shape=obs["rgb_image"].shape, dtype="uint8"),
-                depth_image=spaces.Box(0, self.depth_range[1], shape=obs["depth_image"].shape, dtype="float64")
+                depth_image=spaces.Box(0, self.depth_range[1], shape=obs["depth_image"].shape, dtype="float32")
             )
         )
